[SQL] fix edge case for time arithmetic with interval
