+++ sql-to-dbsp-compiler/SQL-compiler/src/test/java/org/dbsp/sqlCompiler/compiler/sql/functions/FunctionsTest.java
@@ -72,8 +72,7 @@ public class FunctionsTest extends SqlIoTest {
     }
 
 
-    @Test
-    @Ignore("https://issues.apache.org/jira/projects/CALCITE/issues/CALCITE-6210")
+    @Test @Ignore("https://issues.apache.org/jira/projects/CALCITE/issues/CALCITE-6210")
     public void testBinaryCast() {
         this.qs(
                 """
@@ -82,7 +81,7 @@ public class FunctionsTest extends SqlIoTest {
                         -----
                          34567890
                         (1 row)
-                                                
+                        
                         SELECT '1234567890'::VARBINARY as val;
                          val
                         -----
@@ -115,115 +114,115 @@ public class FunctionsTest extends SqlIoTest {
                         ---
                          0
                         (1 row)
-                                                
+                        
                         SELECT (-32768)::int2 % (-1)::int2 as x;
                         x
                         ---
                          0
                         (1 row)
-                                                
+                        
                         SELECT (-2147483648)::int4 % (-1)::int4 as x;
                         x
                         ---
                          0
                         (1 row)
-                                                
+                        
                         SELECT (-9223372036854775808)::int64 % (-1)::int64 as x;
                         x
                         ---
                          0
                         (1 row)
-                                                
+                        
                         SELECT (8)::tinyint % (3)::tinyint as x;
                         x
                         ---
                          2
                         (1 row)
-                                                
+                        
                         SELECT (8)::int2 % (3)::int2 as x;
                         x
                         ---
                          2
                         (1 row)
-                                                
+                        
                         SELECT (8)::int4 % (3)::int4 as x;
                         x
                         ---
                          2
                         (1 row)
-                                                
+                        
                         SELECT (8)::int64 % (3)::int64 as x;
                         x
                         ---
                          2
                         (1 row)
-                                                
+                        
                         SELECT (8)::tinyint % (-3)::tinyint as x;
                         x
                         ---
                          2
                         (1 row)
-                                                
+                        
                         SELECT (8)::int2 % (-3)::int2 as x;
                         x
                         ---
                          2
                         (1 row)
-                                                
+                        
                         SELECT (8)::int4 % (-3)::int4 as x;
                         x
                         ---
                          2
                         (1 row)
-                                                
+                        
                         SELECT (8)::int64 % (-3)::int64 as x;
                         x
                         ---
                          2
                         (1 row)
-                                                
+                        
                         SELECT (-8)::tinyint % (3)::tinyint as x;
                         x
                         ---
                          -2
                         (1 row)
-                                                
+                        
                         SELECT (-8)::int2 % (3)::int2 as x;
                         x
                         ---
                          -2
                         (1 row)
-                                                
+                        
                         SELECT (-8)::int4 % (3)::int4 as x;
                         x
                         ---
                          -2
                         (1 row)
-                                                
+                        
                         SELECT (-8)::int64 % (3)::int64 as x;
                         x
                         ---
                          -2
                         (1 row)
-                                                
+                        
                         SELECT (-8)::tinyint % (-3)::tinyint as x;
                         x
                         ---
                          -2
                         (1 row)
-                                                
+                        
                         SELECT (-8)::int2 % (-3)::int2 as x;
                         x
                         ---
                          -2
                         (1 row)
-                                                
+                        
                         SELECT (-8)::int4 % (-3)::int4 as x;
                         x
                         ---
                          -2
                         (1 row)
-                                                
+                        
                         SELECT (-8)::int64 % (-3)::int64 as x;
                         x
                         ---
@@ -300,116 +299,116 @@ public class FunctionsTest extends SqlIoTest {
                 ------
                  1234.13
                 (1 row)
-                                
+                
                 select CAST((CAST('1234.1234' AS DECIMAL(8, 4))) AS DECIMAL(6, 2));
                  cast
                 ------
                  1234.12
                 (1 row)
-                                
+                
                 select CAST((CAST('-1234.1264' AS DECIMAL(8, 4))) AS DECIMAL(6, 2));
                  cast
                 ------
                  -1234.13
                 (1 row)
-                                
+                
                 select CAST((CAST('-1234.1234' AS DECIMAL(8, 4))) AS DECIMAL(6, 2));
                  cast
                 ------
                  -1234.12
                 (1 row)
-                                
+                
                 select CAST((CAST('1234.1250' AS DECIMAL(8, 4))) AS DECIMAL(6, 2));
                  cast
                 ------
                  1234.13
                 (1 row)
-                                
+                
                 select CAST((CAST('-1234.1250' AS DECIMAL(8, 4))) AS DECIMAL(6, 2));
                  cast
                 ------
                  -1234.13
                 (1 row)
-                                
+                
                 select CAST((CAST('.1264' AS DECIMAL(4, 4))) AS DECIMAL(2, 2));
                  cast
                 ------
                  0.13
                 (1 row)
-                                
+                
                 select CAST((CAST('.1234' AS DECIMAL(4, 4))) AS DECIMAL(2, 2));
                  cast
                 ------
                  0.12
                 (1 row)
-                                
+                
                 select CAST((CAST('-.1264' AS DECIMAL(4, 4))) AS DECIMAL(2, 2));
                  cast
                 ------
                  -0.13
                 (1 row)
-                                
+                
                 select CAST((CAST('-.1234' AS DECIMAL(4, 4))) AS DECIMAL(2, 2));
                  cast
                 ------
                  -0.12
                 (1 row)
-                                
+                
                 select CAST((CAST('00.1264' AS DECIMAL(4, 4))) AS DECIMAL(2, 2));
                  cast
                 ------
                  0.13
                 (1 row)
-                                
+                
                 select CAST((CAST('00.1234' AS DECIMAL(4, 4))) AS DECIMAL(2, 2));
                  cast
                 ------
                  0.12
                 (1 row)
-                                
+                
                 select CAST((CAST('-00.1264' AS DECIMAL(4, 4))) AS DECIMAL(2, 2));
                  cast
                 ------
                  -0.13
                 (1 row)
-                                
+                
                 select CAST((CAST('-00.1234' AS DECIMAL(4, 4))) AS DECIMAL(2, 2));
                  cast
                 ------
                  -0.12
                 (1 row)
-                                
+                
                 -- the following tests are from mysql
                 select cast('1.00000001335143196001808973960578441619873046875E-10' as decimal(30,15));
                      decimal
                 -------------------
                  0.000000000100000
                 (1 row)
-                                
+                
                 select ln(14000) as c1, cast(ln(14000) as decimal(5,3)) as c2, cast(ln(14000) as decimal(5,3)) as c3;
                         c1         |  c2   |  c3
                 -------------------+-------+-------
                  9.546812608597396 | 9.547 | 9.547
                 (1 row)
-                                
+                
                 select cast(143.481 as decimal(4,1));
                  cast(143.481 as decimal(4,1))
                 -------------------------------
                 143.5
                 (1 row)
-                                
+                
                 select cast(143.481 as decimal(4,0));
                  cast(143.481 as decimal(4,0))
                 -------------------------------
                 143
                 (1 row)
-                                
+                
                 select cast(-3.4 as decimal(2,1));
                  cast(-3.4 as decimal(2,1))
                 -------------------------------
                 -3.4
                 (1 row)
-                                
+                
                 select cast(98.6 as decimal(2,0));
                  cast(98.6 as decimal(2,0))
                 -------------------------------
@@ -454,7 +453,7 @@ public class FunctionsTest extends SqlIoTest {
                 ----
                  0.693147180559945
                 (1 row)
-                                
+                
                 SELECT ln(2.0);
                  ln
                 ----
@@ -473,57 +472,57 @@ public class FunctionsTest extends SqlIoTest {
                         -------
                         null
                         (1 row)
-                                                
+                        
                         SELECT IS_INF(1);
                         is_inf
                         -------
                          f
                         (1 row)
-                                                
+                        
                         SELECT IS_INF('INF'::DOUBLE);
                         is_inf
                         -------
                          t
                         (1 row)
-                                                
+                        
                         SELECT IS_INF('-INF'::DOUBLE);
                         is_inf
                         -------
                          t
                         (1 row)
-                                                
+                        
                         SELECT IS_INF('Infinity'::DOUBLE);
                         is_inf
                         -------
                          t
                         (1 row)
-                                                
+                        
                         SELECT IS_INF('-Infinity'::DOUBLE);
                         is_inf
                         -------
                          t
                         (1 row)
-                                                
+                        
                         -- f64::MAX
                         SELECT IS_INF(1.7976931348623157e308);
                         is_inf
                         -------
                          f
                         (1 row)
-                                                
+                        
                         -- f64::MIN
                         SELECT IS_INF(-1.7976931348623157e308);
                         is_inf
                         -------
                          f
                         (1 row)
-                                        
+                
                         SELECT IS_INF(1e0 / 0e0);
                          IS_INF
                         --------
                          t
                         (1 row)
-                                        
+                
                         SELECT IS_INF(-1e0 / 0e0);
                          IS_INF
                         --------
@@ -533,8 +532,7 @@ public class FunctionsTest extends SqlIoTest {
         );
     }
 
-    @Test
-    @Ignore("Calcite bug: https://github.com/feldera/feldera/issues/1345")
+    @Test @Ignore("Calcite bug: https://github.com/feldera/feldera/issues/1345")
     public void testIsInfReal() {
         this.qs("""
                 -- f64::MAX
@@ -543,7 +541,7 @@ public class FunctionsTest extends SqlIoTest {
                 -------
                  t
                 (1 row)
-                                
+                
                 -- f64::MIN
                 SELECT is_inf(-1.7976931348623157e308::real);
                 real
@@ -562,37 +560,37 @@ public class FunctionsTest extends SqlIoTest {
                 --------
                  null
                 (1 row)
-                                
+                
                 SELECT is_nan('nan'::double);
                  is_nan
                 --------
                  t
                 (1 row)
-                                
+                
                 SELECT is_nan(0e0 / 0e0);
                  is_nan
                 --------
                  t
                 (1 row)
-                                
+                
                 SELECT is_nan(-0e0 / 0e0);
                  is_nan
                 --------
                  t
                 (1 row)
-                                
+                
                 SELECT is_nan(0.0);
                  is_nan
                 --------
                  f
                 (1 row)
-                                
+                
                 SELECT is_nan(1.7976931348623157e308);
                  is_nan
                 --------
                  f
                 (1 row)
-                                
+                
                 SELECT is_nan(-1.7976931348623157e308);
                  is_nan
                 --------
@@ -610,7 +608,7 @@ public class FunctionsTest extends SqlIoTest {
                 -------
                  4.0000000000000000
                 (1 row)
-                                
+                
                 SELECT power(2e0::real, 2);
                  power
                 -------
@@ -677,7 +675,7 @@ public class FunctionsTest extends SqlIoTest {
                 ------
                  3
                 (1 row)
-                                
+                
                 SELECT sqrt(null::double);
                  sqrt
                 ------
@@ -700,536 +698,536 @@ public class FunctionsTest extends SqlIoTest {
                 ------------
                  15
                 (1 row)
-                                
+                
                 select round(15.4);
                  round(15.4)
                 ------------
                  15
                 (1 row)
-                                
+                
                 select round(15.5);
                  round(15.5)
                 ------------
                  16
                 (1 row)
-                                
+                
                 select round(15.6);
                  round(15.6)
                 ------------
                  16
                 (1 row)
-                                
+                
                 select round(15.9);
                  round(15.9)
                 ------------
                  16
                 (1 row)
-                                
+                
                 select round(-15.1);
                  round(-15.1)
                 ------------
                  -15
                 (1 row)
-                                
+                
                 select round(-15.4);
                  round(-15.4)
                 ------------
                  -15
                 (1 row)
-                                
+                
                 select round(-15.5);
                  round(-15.5)
                 ------------
                  -16
                 (1 row)
-                                
+                
                 select round(-15.6);
                  round(-15.6)
                 ------------
                  -16
                 (1 row)
-                                
+                
                 select round(-15.9);
                  round(-15.9)
                 ------------
                  -16
                 (1 row)
-                                
+                
                 select round(15.1,1);
                  round(15.1,1)
                 ------------
                  15.1
                 (1 row)
-                                
+                
                 select round(15.4,1);
                  round(15.4,1)
                 ------------
                  15.4
                 (1 row)
-                                
+                
                 select round(15.5,1);
                  round(15.5,1)
                 ------------
                  15.5
                 (1 row)
-                                
+                
                 select round(15.6,1);
                  round(15.6,1)
                 ------------
                  15.6
                 (1 row)
-                                
+                
                 select round(15.9,1);
                  round(15.9,1)
                 ------------
                  15.9
                 (1 row)
-                                
+                
                 select round(-15.1,1);
                 round(-15.1,1)
                 ------------
                  -15.1
                 (1 row)
-                                
+                
                 select round(-15.4,1);
                 round(-15.4,1)
                 ------------
                  -15.4
                 (1 row)
-                                
+                
                 select round(-15.5,1);
                 round(-15.5,1)
                 ------------
                  -15.5
                 (1 row)
-                                
+                
                 select round(-15.6,1);
                 round(-15.6,1)
                 ------------
                  -15.6
                 (1 row)
-                                
+                
                 select round(-15.9,1);
                 round(-15.9,1)
                 ------------
                  -15.9
                 (1 row)
-                                
+                
                 select round(15.1,0);
                 round(15.1,0)
                 ------------
                  15
                 (1 row)
-                                
+                
                 select round(15.4,0);
                  round(15.4,0)
                 ------------
                  15
                 (1 row)
-                                
+                
                 select round(15.5,0);
                 round(15.5,0)
                 ------------
                  16
                 (1 row)
-                                
+                
                 select round(15.6,0);
                 round(15.6,0)
                 ------------
                  16
                 (1 row)
-                                
+                
                 select round(15.9,0);
                 round(15.9,0)
                 ------------
                  16
                 (1 row)
-                                
+                
                 select round(-15.1,0);
                 round(-15.1,0)
                 ------------
                  -15
                 (1 row)
-                                
+                
                 select round(-15.4,0);
                 round(-15.4,0)
                 ------------
                  -15
                 (1 row)
-                                
+                
                 select round(-15.5,0);
                 round(-15.5,0)
                 ------------
                  -16
                 (1 row)
-                                
+                
                 select round(-15.6,0);
                 round(-15.6,0)
                 ------------
                  -16
                 (1 row)
-                                
+                
                 select round(-15.9,0);
                 round(-15.9,0)
                 ------------
                  -16
                 (1 row)
-                                
+                
                 select round(15.1,-1);
                 round(15.1,-1)
                 ------------
                  20
                 (1 row)
-                                
+                
                 select round(15.4,-1);
                 round(15.4,-1)
                 ------------
                  20
                 (1 row)
-                                
+                
                 select round(15.5,-1);
                 round(15.5,-1)
                 ------------
                  20
                 (1 row)
-                                
+                
                 select round(15.6,-1);
                 round(15.6,-1)
                 ------------
                  20
                 (1 row)
-                                
+                
                 select round(15.9,-1);
                 round(15.9,-1)
                 ------------
                  20
                 (1 row)
-                                
+                
                 select round(-15.1,-1);
                 round(-15.1,-1)
                 ------------
                  -20
                 (1 row)
-                                
+                
                 select round(-15.4,-1);
                 round(-15.4,-1)
                 ------------
                  -20
                 (1 row)
-                                
+                
                 select round(-15.5,-1);
                 round(-15.5,-1)
                 ------------
                  -20
                 (1 row)
-                                
+                
                 select round(-15.6,-1);
                 round(-15.6,-1)
                 ------------
                  -20
                 (1 row)
-                                
+                
                 select round(-15.91,-1);
                 round(-15.91,-1)
                 ------------
                  -20
                 (1 row)
-                                
+                
                 select round(-15.91,-1::tinyint);
                 round(-15.91,-1)
                 ------------
                  -20
                 (1 row)
-                                
+                
                 select round(-15.91,-1::smallint);
                 round(-15.91,-1)
                 ------------
                  -20
                 (1 row)
-                                
+                
                 select round(-15.91,-1::int);
                 round(-15.91,-1)
                 ------------
                  -20
                 (1 row)
-                                
-                                
+                
+                
                 select round(15.1);
                  round(15.1)
                 ------------
                  15
                 (1 row)
-                                
+                
                 select round(15.4);
                  round(15.4)
                 ------------
                  15
                 (1 row)
-                                
+                
                 select round(15.5);
                  round(15.5)
                 ------------
                  16
                 (1 row)
-                                
+                
                 select round(15.6);
                  round(15.6)
                 ------------
                  16
                 (1 row)
-                                
+                
                 select round(15.9);
                  round(15.9)
                 ------------
                  16
                 (1 row)
-                                
+                
                 select round(-15.1);
                  round(-15.1)
                 ------------
                  -15
                 (1 row)
-                                
+                
                 select round(-15.4);
                  round(-15.4)
                 ------------
                  -15
                 (1 row)
-                                
+                
                 select round(-15.5);
                  round(-15.5)
                 ------------
                  -16
                 (1 row)
-                                
+                
                 select round(-15.6);
                  round(-15.6)
                 ------------
                  -16
                 (1 row)
-                                
+                
                 select round(-15.9);
                  round(-15.9)
                 ------------
                  -16
                 (1 row)
-                                
+                
                 select round(15.1::DOUBLE,1);
                  round(15.1::DOUBLE,1)
                 ------------
                  15.1
                 (1 row)
-                                
+                
                 select round(15.4::DOUBLE,1);
                  round(15.4::DOUBLE,1)
                 ------------
                  15.4
                 (1 row)
-                                
+                
                 select round(15.5::DOUBLE,1);
                  round(15.5::DOUBLE,1)
                 ------------
                  15.5
                 (1 row)
-                                
+                
                 select round(15.6::DOUBLE,1);
                  round(15.6::DOUBLE,1)
                 ------------
                  15.6
                 (1 row)
-                                
+                
                 select round(15.9::DOUBLE,1);
                  round(15.9::DOUBLE,1)
                 ------------
                  15.9
                 (1 row)
-                                
+                
                 select round(-15.1::DOUBLE,1);
                 round(-15.1::DOUBLE,1)
                 ------------
                  -15.1
                 (1 row)
-                                
+                
                 select round(-15.4::DOUBLE,1);
                 round(-15.4::DOUBLE,1)
                 ------------
                  -15.4
                 (1 row)
-                                
+                
                 select round(-15.5::DOUBLE,1);
                 round(-15.5::DOUBLE,1)
                 ------------
                  -15.5
                 (1 row)
-                                
+                
                 select round(-15.6::DOUBLE,1);
                 round(-15.6::DOUBLE,1)
                 ------------
                  -15.6
                 (1 row)
-                                
+                
                 select round(-15.9::DOUBLE,1);
                 round(-15.9::DOUBLE,1)
                 ------------
                  -15.9
                 (1 row)
-                                
+                
                 select round(15.1::DOUBLE,0);
                 round(15.1::DOUBLE,0)
                 ------------
                  15
                 (1 row)
-                                
+                
                 select round(15.4::DOUBLE,0);
                  round(15.4::DOUBLE,0)
                 ------------
                  15
                 (1 row)
-                                
+                
                 select round(15.5::DOUBLE,0);
                 round(15.5::DOUBLE,0)
                 ------------
                  16
                 (1 row)
-                                
+                
                 select round(15.6::DOUBLE,0);
                 round(15.6::DOUBLE,0)
                 ------------
                  16
                 (1 row)
-                                
+                
                 select round(15.9::DOUBLE,0);
                 round(15.9::DOUBLE,0)
                 ------------
                  16
                 (1 row)
-                                
+                
                 select round(-15.1::DOUBLE,0);
                 round(-15.1::DOUBLE,0)
                 ------------
                  -15
                 (1 row)
-                                
+                
                 select round(-15.4::DOUBLE,0);
                 round(-15.4::DOUBLE,0)
                 ------------
                  -15
                 (1 row)
-                                
+                
                 select round(-15.5::DOUBLE,0);
                 round(-15.5::DOUBLE,0)
                 ------------
                  -16
                 (1 row)
-                                
+                
                 select round(-15.6::DOUBLE,0);
                 round(-15.6::DOUBLE,0)
                 ------------
                  -16
                 (1 row)
-                                
+                
                 select round(-15.9::DOUBLE,0);
                 round(-15.9::DOUBLE,0)
                 ------------
                  -16
                 (1 row)
-                                
+                
                 select round(15.1::DOUBLE,-1);
                 round(15.1::DOUBLE,-1)
                 ------------
                  20
                 (1 row)
-                                
+                
                 select round(15.4::DOUBLE,-1);
                 round(15.4::DOUBLE,-1)
                 ------------
                  20
                 (1 row)
-                                
+                
                 select round(15.5::DOUBLE,-1);
                 round(15.5::DOUBLE,-1)
                 ------------
                  20
                 (1 row)
-                                
+                
                 select round(15.6::DOUBLE,-1);
                 round(15.6::DOUBLE,-1)
                 ------------
                  20
                 (1 row)
-                                
+                
                 select round(15.9::DOUBLE,-1);
                 round(15.9::DOUBLE,-1)
                 ------------
                  20
                 (1 row)
-                                
+                
                 select round(-15.1::DOUBLE,-1);
                 round(-15.1::DOUBLE,-1)
                 ------------
                  -20
                 (1 row)
-                                
+                
                 select round(-15.4::DOUBLE,-1);
                 round(-15.4::DOUBLE,-1)
                 ------------
                  -20
                 (1 row)
-                                
+                
                 select round(-15.5::DOUBLE,-1);
                 round(-15.5::DOUBLE,-1)
                 ------------
                  -20
                 (1 row)
-                                
+                
                 select round(-15.6::DOUBLE,-1);
                 round(-15.6::DOUBLE,-1)
                 ------------
                  -20
                 (1 row)
-                                
+                
                 select round(-15.91::DOUBLE,-1);
                 round(-15.91::DOUBLE,-1)
                 ------------
                  -20
                 (1 row)
-                                
+                
                 select round(-15.91::DOUBLE,-1::tinyint);
                 round(-15.91::DOUBLE,-1)
                 ------------
                  -20
                 (1 row)
-                                
+                
                 select round(-15.91::DOUBLE,-1::smallint);
                 round(-15.91::DOUBLE,-1)
                 ------------
                  -20
                 (1 row)
-                                
+                
                 select round(-15.91::DOUBLE,-1::int);
                 round(-15.91::DOUBLE,-1)
                 ------------
                  -20
                 (1 row)
-                                
+                
                 select round(null::double);
                  round
                 -------
                  NULL
                 (1 row)
-                                
+                
                 select round(null::double, null);
                  round
                 -------
                  NULL
                 (1 row)
-                                
+                
                 select round(null::DECIMAL);
                  round
                 -------
                  NULL
                 (1 row)
-                                
+                
                 select round(null::DECIMAL, null);
                  round
                 -------
@@ -1247,314 +1245,314 @@ public class FunctionsTest extends SqlIoTest {
                 -----
                 5678
                 (1 row)
-                                
+                
                 select truncate(5678.123451,0);
                 truncate(5678.123451,0)
                 -----
                 5678
                 (1 row)
-                                
+                
                 select truncate(5678.123451,1);
                 truncate(5678.123451,1)
                 -----
                 5678.1
                 (1 row)
-                                
+                
                 select truncate(5678.123451,2);
                 truncate(5678.123451,2)
                 -----
                 5678.12
                 (1 row)
-                                
+                
                 select truncate(5678.123451,3);
                 truncate(5678.123451,3)
                 -----
                 5678.123
                 (1 row)
-                                
+                
                 select truncate(5678.123451,4);
                 truncate(5678.123451,4)
                 -----
                 5678.1234
                 (1 row)
-                                
+                
                 select truncate(5678.123451,5);
                 truncate(5678.123451,5)
                 -----
                 5678.12345
                 (1 row)
-                                
+                
                 select truncate(5678.123451,6);
                 truncate(5678.123451,6)
                 -----
                 5678.123451
                 (1 row)
-                                
+                
                 select truncate(5678.123451,-1);
                 truncate(5678.123451,-1)
                 -----
                 5670
                 (1 row)
-                                
+                
                 select truncate(5678.123451,-2);
                 truncate(5678.123451,-2)
                 -----
                 5600
                 (1 row)
-                                
+                
                 select truncate(5678.123451,-3);
                 truncate(5678.123451,-3)
                 -----
                 5000
                 (1 row)
-                                
+                
                 select truncate(5678.123451,-4);
                 truncate(5678.123451,-4)
                 -----
                 0
                 (1 row)
-                                
+                
                 select truncate(-5678.123451,0);
                 truncate(-5678.123451,0)
                 -----
                 -5678
                 (1 row)
-                                
+                
                 select truncate(-5678.123451,1);
                 truncate(-5678.123451,1)
                 -----
                 -5678.1
                 (1 row)
-                                
+                
                 select truncate(-5678.123451,2);
                 truncate(-5678.123451,2)
                 -----
                 -5678.12
                 (1 row)
-                                
+                
                 select truncate(-5678.123451,3);
                 truncate(-5678.123451,3)
                 -----
                 -5678.123
                 (1 row)
-                                
+                
                 select truncate(-5678.123451,4);
                 truncate(-5678.123451,4)
                 -----
                 -5678.1234
                 (1 row)
-                                
+                
                 select truncate(-5678.123451,5);
                 truncate(-5678.123451,5)
                 -----
                 -5678.12345
                 (1 row)
-                                
+                
                 select truncate(-5678.123451,6);
                 truncate(-5678.123451,6)
                 -----
                 -5678.123451
                 (1 row)
-                                
+                
                 select truncate(-5678.123451,-1);
                 truncate(-5678.123451,-1)
                 -----
                 -5670
                 (1 row)
-                                
+                
                 select truncate(-5678.123451,-2);
                 truncate(-5678.123451,-2)
                 -----
                 -5600
                 (1 row)
-                                
+                
                 select truncate(-5678.123451,-3);
                 truncate(-5678.123451,-3)
                 -----
                 -5000
                 (1 row)
-                                
+                
                 select truncate(-5678.123451,-4);
                 truncate(-5678.123451,-4)
                 -----
                 0
                 (1 row)
-                                
+                
                 select truncate(5678.123451,1::tinyint);
                 truncate(5678.123451,1)
                 -----
                 5678.1
                 (1 row)
-                                
+                
                 select truncate(5678.123451,1::smallint);
                 truncate(5678.123451,1)
                 -----
                 5678.1
                 (1 row)
-                                
+                
                 select truncate(5678.123451,1::int);
                 truncate(5678.123451,1)
                 -----
                 5678.1
                 (1 row)
-                                
-                                
+                
+                
                 select truncate(5678.123451);
                 truncate(5678.123451)
                 -----
                 5678
                 (1 row)
-                                
+                
                 select truncate(5678.123451::DOUBLE,0);
                 truncate(5678.123451::DOUBLE,0)
                 -----
                 5678
                 (1 row)
-                                
+                
                 select truncate(5678.123451::DOUBLE,1);
                 truncate(5678.123451::DOUBLE,1)
                 -----
                 5678.1
                 (1 row)
-                                
+                
                 select truncate(5678.123451::DOUBLE,2);
                 truncate(5678.123451::DOUBLE,2)
                 -----
                 5678.12
                 (1 row)
-                                
+                
                 select truncate(5678.123451::DOUBLE,3);
                 truncate(5678.123451::DOUBLE,3)
                 -----
                 5678.123
                 (1 row)
-                                
+                
                 select truncate(5678.123451::DOUBLE,4);
                 truncate(5678.123451::DOUBLE,4)
                 -----
                 5678.1234
                 (1 row)
-                                
+                
                 select truncate(5678.123451::DOUBLE,5);
                 truncate(5678.123451::DOUBLE,5)
                 -----
                 5678.12345
                 (1 row)
-                                
+                
                 select truncate(5678.123451::DOUBLE,6);
                 truncate(5678.123451::DOUBLE,6)
                 -----
                 5678.123451
                 (1 row)
-                                
+                
                 select truncate(5678.123451::DOUBLE,-1);
                 truncate(5678.123451::DOUBLE,-1)
                 -----
                 5670
                 (1 row)
-                                
+                
                 select truncate(5678.123451::DOUBLE,-2);
                 truncate(5678.123451::DOUBLE,-2)
                 -----
                 5600
                 (1 row)
-                                
+                
                 select truncate(5678.123451::DOUBLE,-3);
                 truncate(5678.123451::DOUBLE,-3)
                 -----
                 5000
                 (1 row)
-                                
+                
                 select truncate(5678.123451::DOUBLE,-4);
                 truncate(5678.123451::DOUBLE,-4)
                 -----
                 0
                 (1 row)
-                                
+                
                 select truncate(-5678.123451::DOUBLE,0);
                 truncate(-5678.123451::DOUBLE,0)
                 -----
                 -5678
                 (1 row)
-                                
+                
                 select truncate(-5678.123451::DOUBLE,1);
                 truncate(-5678.123451::DOUBLE,1)
                 -----
                 -5678.1
                 (1 row)
-                                
+                
                 select truncate(-5678.123451::DOUBLE,2);
                 truncate(-5678.123451::DOUBLE,2)
                 -----
                 -5678.12
                 (1 row)
-                                
+                
                 select truncate(-5678.123451::DOUBLE,3);
                 truncate(-5678.123451::DOUBLE,3)
                 -----
                 -5678.123
                 (1 row)
-                                
+                
                 select truncate(-5678.123451::DOUBLE,4);
                 truncate(-5678.123451::DOUBLE,4)
                 -----
                 -5678.1234
                 (1 row)
-                                
+                
                 select truncate(-5678.123451::DOUBLE,5);
                 truncate(-5678.123451::DOUBLE,5)
                 -----
                 -5678.12345
                 (1 row)
-                                
+                
                 select truncate(-5678.123451::DOUBLE,6);
                 truncate(-5678.123451::DOUBLE,6)
                 -----
                 -5678.123451
                 (1 row)
-                                
+                
                 select truncate(-5678.123451::DOUBLE,-1);
                 truncate(-5678.123451::DOUBLE,-1)
                 -----
                 -5670
                 (1 row)
-                                
+                
                 select truncate(-5678.123451::DOUBLE,-2);
                 truncate(-5678.123451::DOUBLE,-2)
                 -----
                 -5600
                 (1 row)
-                                
+                
                 select truncate(-5678.123451::DOUBLE,-3);
                 truncate(-5678.123451::DOUBLE,-3)
                 -----
                 -5000
                 (1 row)
-                                
+                
                 select truncate(-5678.123451::DOUBLE,-4);
                 truncate(-5678.123451::DOUBLE,-4)
                 -----
                 0
                 (1 row)
-                                
+                
                 select truncate(5678.123451::DOUBLE,1::tinyint);
                 truncate(5678.123451::DOUBLE,1)
                 -----
                 5678.1
                 (1 row)
-                                
+                
                 select truncate(5678.123451::DOUBLE,1::smallint);
                 truncate(5678.123451::DOUBLE,1)
                 -----
                 5678.1
                 (1 row)
-                                
+                
                 select truncate(5678.123451::DOUBLE,1::int);
                 truncate(5678.123451::DOUBLE,1)
                 -----
                 5678.1
                 (1 row)
-                                
+                
                 """
         );
     }
@@ -1568,31 +1566,31 @@ public class FunctionsTest extends SqlIoTest {
                 ----------
                  0.00
                 (1 row)
-                                
+                
                 select truncate(-0.00123::DECIMAL(10, 5), 2);
                  truncate
                 ----------
                  0.00
                 (1 row)
-                                
+                
                 select round(-0.00123::DOUBLE, 2);
                  round
                 ----------
                  0.00
                 (1 row)
-                                
+                
                 select round(-0.00123::DECIMAL(10, 5), 2);
                  round
                 ----------
                  0.00
                 (1 row)
-                                
+                
                 select round(-0.00623::DOUBLE, 2);
                  round
                 ----------
                  -0.01
                 (1 row)
-                                
+                
                 select round(-0.00623::DECIMAL(10, 5), 2);
                  round
                 ----------
@@ -1602,8 +1600,7 @@ public class FunctionsTest extends SqlIoTest {
         );
     }
 
-    @Test
-    @Ignore("https://github.com/feldera/feldera/issues/1379")
+    @Test @Ignore("https://github.com/feldera/feldera/issues/1379")
     public void testRoundBigInt() {
         this.q("""
                 SELECT round(123.123, 2::bigint);
@@ -1613,8 +1610,7 @@ public class FunctionsTest extends SqlIoTest {
         );
     }
 
-    @Test
-    @Ignore("https://github.com/feldera/feldera/issues/1379")
+    @Test @Ignore("https://github.com/feldera/feldera/issues/1379")
     public void testTruncateBigInt() {
         this.q("""
                 select truncate(5678.123451,1::bigint);
@@ -1632,25 +1628,25 @@ public class FunctionsTest extends SqlIoTest {
                 -----
                  1.0
                 (1 row)
-                                
+                
                 SELECT EXP(0e0);
                  exp
                 -----
                  1.0
                 (1 row)
-                                
+                
                 SELECT EXP(null);
                  exp
                 -----
                 null
                 (1 row)
-                                
+                
                 SELECT EXP(null::double);
                  exp
                 -----
                 null
                 (1 row)
-                                
+                
                 SELECT EXP(0);
                  exp
                 -----
@@ -1660,8 +1656,7 @@ public class FunctionsTest extends SqlIoTest {
         );
     }
 
-    @Test
-    @Ignore("FP comparison error for Calcite optimized version")
+    @Test @Ignore("FP comparison error for Calcite optimized version")
     public void testExpEdgeCase() {
         this.qs("""
                 -- changed the type from numeric to double
@@ -1684,25 +1679,25 @@ public class FunctionsTest extends SqlIoTest {
                 ------------
                  feldera
                 (1 row)
-                                
+                
                 SELECT GUNZIP(x'1f8b08000000000000ff734bcd49492d4a0400bdb8a86307000000'::bytea);
                  gunzip
                 ------------
                  Feldera
                 (1 row)
-                                
+                
                 SELECT GUNZIP(x'1f8b08000000000000ffcb48cdc9c9070086a6103605000000'::bytea);
                  gunzip
                 ------------
                  hello
                 (1 row)
-                                
+                
                 SELECT GUNZIP(x'1f8b08000000000000132bc9c82c5600a2dc4a851282ccd48a12002e7a22ff30000000'::bytea);
                  gunzip
                 --------------------------------------------------------------------------
                  this is my this is my this is my this is my text
                 (1 row)
-                                
+                
                 SELECT GUNZIP(null);
                  gunzip
                 --------
@@ -1725,19 +1720,19 @@ public class FunctionsTest extends SqlIoTest {
                 -------
                  123.12
                 (1 row)
-                                
+                
                 SELECT ROUND(123.1266::DOUBLE, 2);
                  round
                 -------
                  123.13
                 (1 row)
-                                
+                
                 SELECT ROUND(123.1266::DOUBLE, -2);
                  round
                 -------
                  100
                 (1 row)
-                                
+                
                 SELECT ROUND(123.1266::DOUBLE);
                  round
                 -------
@@ -1755,19 +1750,19 @@ public class FunctionsTest extends SqlIoTest {
                 -------
                  true
                 (1 row)
-                                
+                
                 SELECT RLIKE(null, 's..i.*');
                  rlike
                 -------
                  NULL
                 (1 row)
-                                
+                
                 SELECT RLIKE('string', null);
                  rlike
                 -------
                  NULL
                 (1 row)
-                                
+                
                 SELECT RLIKE(null, null);
                  rlike
                 -------
@@ -1785,73 +1780,73 @@ public class FunctionsTest extends SqlIoTest {
                 ----------
                  {1, 2, 3, 4, 5, 6, 7, 8, 9, 10}
                 (1 row)
-                                
+                
                 SELECT SEQUENCE(10, 1);
                  sequence
                 ----------
                  {}
                 (1 row)
-                                
+                
                 SELECT SEQUENCE(1, 1);
                  sequence
                 ----------
                  {1}
                 (1 row)
-                                
+                
                 SELECT SEQUENCE(-6, 1);
                  sequence
                 ----------
                  {-6, -5, -4, -3, -2, -1, 0, 1}
                 (1 row)
-                                
+                
                 SELECT SEQUENCE(-6, -9);
                  sequence
                 ----------
                  {}
                 (1 row)
-                                
+                
                 SELECT SEQUENCE(0, -2);
                  sequence
                 ----------
                  {}
                 (1 row)
-                                
+                
                 SELECT SEQUENCE(null::int, 1);
                  sequence
                 ----------
                  NULL
                 (1 row)
-                                
+                
                 SELECT SEQUENCE(null, 1);
                  sequence
                 ----------
                  NULL
                 (1 row)
-                                
+                
                 SELECT SEQUENCE(1, null);
                  sequence
                 ----------
                  NULL
                 (1 row)
-                                
+                
                 SELECT SEQUENCE(null, null);
                  sequence
                 ----------
                  NULL
                 (1 row)
-                                
+                
                 SELECT SEQUENCE(null::tinyint, null::tinyint);
                  sequence
                 ----------
                  NULL
                 (1 row)
-                                
+                
                 SELECT SEQUENCE(1::tinyint, 10::tinyint);
                  sequence
                 ----------
                  {1, 2, 3, 4, 5, 6, 7, 8, 9, 10}
                 (1 row)
-                                
+                
                 SELECT SEQUENCE(1::tinyint, 3::bigint);
                  sequence
                 ----------
@@ -1869,49 +1864,49 @@ public class FunctionsTest extends SqlIoTest {
                 ------
                      2
                 (1 row)
-                                
+                
                 SELECT ceil(2::smallint);
                  ceil
                 ------
                      2
                 (1 row)
-                                
+                
                 SELECT ceil(2::int);
                  ceil
                 ------
                      2
                 (1 row)
-                                
+                
                 SELECT ceil(2::bigint);
                  ceil
                 ------
                      2
                 (1 row)
-                                
+                
                 SELECT ceil(null::tinyint);
                  ceil
                 ------
                  NULL
                 (1 row)
-                                
+                
                 SELECT ceil(null::smallint);
                  ceil
                 ------
                  NULL
                 (1 row)
-                                
+                
                 SELECT ceil(null::int);
                  ceil
                 ------
                  NULL
                 (1 row)
-                                
+                
                 SELECT ceil(null::bigint);
                  ceil
                 ------
                  NULL
                 (1 row)
-                                
+                
                 SELECT ceil(null);
                  ceil
                 ------
@@ -1929,49 +1924,49 @@ public class FunctionsTest extends SqlIoTest {
                 ------
                      2
                 (1 row)
-                                
+                
                 SELECT floor(2::smallint);
                  floor
                 ------
                      2
                 (1 row)
-                                
+                
                 SELECT floor(2::int);
                  floor
                 ------
                      2
                 (1 row)
-                                
+                
                 SELECT floor(2::bigint);
                  floor
                 ------
                      2
                 (1 row)
-                                
+                
                 SELECT floor(null::tinyint);
                  floor
                 ------
                  NULL
                 (1 row)
-                                
+                
                 SELECT floor(null::smallint);
                  floor
                 ------
                  NULL
                 (1 row)
-                                
+                
                 SELECT floor(null::int);
                  floor
                 ------
                  NULL
                 (1 row)
-                                
+                
                 SELECT floor(null::bigint);
                  floor
                 ------
                  NULL
                 (1 row)
-                                
+                
                 SELECT floor(null);
                  floor
                 ------
@@ -1979,315 +1974,5 @@ public class FunctionsTest extends SqlIoTest {
                 (1 row)
                 """
         );
-    }
-
-    @Test
-    public void testTimeAddInterval() {
-        this.qs("""
-                SELECT '23:00:00'::time + INTERVAL '10' MINUTES;
-                   time
-                ----------
-                 23:10:00
-                (1 row)
-                
-                SELECT '00:00:00'::time + INTERVAL '10' MINUTES;
-                   time
-                --------
-                 00:10:00
-                (1 row)
-                                
-                SELECT '00:00:00'::time + INTERVAL '-10' MINUTES;
-                   time
-                --------
-                 23:50:00
-                (1 row)
-                                
-                SELECT '00:00:00'::time + INTERVAL '2' DAYS;
-                 time
-                --------
-                 00:00:00
-                (1 row)
-                                
-                SELECT '00:00:00'::time + INTERVAL '-2' DAYS;
-                 time
-                --------
-                 00:00:00
-                (1 row)
-                                
-                SELECT '00:00:00'::time + INTERVAL '2' HOURS;
-                 time
-                --------
-                 02:00:00
-                (1 row)
-                                
-                SELECT '00:00:00'::time + INTERVAL '-2' HOURS;
-                 time
-                --------
-                 22:00:00
-                (1 row)
-                                
-                SELECT '00:00:00'::time + INTERVAL '2' SECONDS;
-                 time
-                --------
-                 00:00:02
-                (1 row)
-                                
-                SELECT '00:00:00'::time + INTERVAL '-2' SECONDS;
-                 time
-                --------
-                 23:59:58
-                (1 row)
-                                
-                SELECT '00:00:00'::time + INTERVAL '2' HOURS + INTERVAL '10' MINUTES + INTERVAL '10' SECONDS;
-                 time
-                --------
-                 02:10:10
-                (1 row)
-                                
-                SELECT '00:00:00'::time + INTERVAL '-2' HOURS + INTERVAL '-10' MINUTES + INTERVAL '-10' SECONDS;
-                 time
-                --------
-                 21:49:50
-                (1 row)
-                                
-                SELECT '00:00:00'::time + INTERVAL '2' HOURS + INTERVAL '10' MINUTES + INTERVAL '-10' SECONDS;
-                 time
-                --------
-                 02:09:50
-                (1 row)
-                                
-                SELECT '00:00:00'::time + INTERVAL '-2' HOURS + INTERVAL '-10' MINUTES + INTERVAL '10' SECONDS;
-                 time
-                --------
-                 21:50:10
-                (1 row)
-                                
-                SELECT '00:00:00'::time + INTERVAL '2' HOURS + INTERVAL '-10' MINUTES + INTERVAL '10' SECONDS;
-                 time
-                --------
-                 01:50:10
-                (1 row)
-                                
-                SELECT '00:00:00'::time + INTERVAL '-2' HOURS + INTERVAL '10' MINUTES + INTERVAL '-10' SECONDS;
-                 time
-                --------
-                 22:09:50
-                (1 row)
-                                
-                SELECT '00:00:00'::time + INTERVAL '2' HOURS + INTERVAL '-10' MINUTES + INTERVAL '-10' SECONDS;
-                 time
-                --------
-                 01:49:50
-                (1 row)
-                                
-                SELECT '00:00:00'::time + INTERVAL '-2' HOURS + INTERVAL '10' MINUTES + INTERVAL '10' SECONDS;
-                 time
-                --------
-                 22:10:10
-                (1 row)
-                                
-                SELECT '23:59:59'::time + INTERVAL '1' SECONDS;
-                   time
-                ----------
-                 00:00:00
-                (1 row)
-                                
-                SELECT '23:59:59'::time + INTERVAL '1' HOURS;
-                   time
-                --------
-                 00:59:59
-                (1 row)
-                                
-                SELECT '23:59:59'::time + INTERVAL '1' MINUTES;
-                   time
-                --------
-                 00:00:59
-                (1 row)
-                                
-                SELECT '23:59:59'::time + INTERVAL '1' DAYS;
-                   time
-                --------
-                 23:59:59
-                (1 row)
-                                
-                SELECT '23:59:59'::time + INTERVAL '1' HOURS + INTERVAL '1' MINUTES + INTERVAL '1' SECONDS;
-                   time
-                --------
-                 01:01:00
-                (1 row)
-                                
-                SELECT '23:59:59'::time + INTERVAL '1' HOURS + INTERVAL '1' MINUTES + INTERVAL '1' SECONDS + INTERVAL '1' DAYS;
-                   time
-                --------
-                 01:01:00
-                (1 row)
-                                
-                SELECT '23:59:59'::time + INTERVAL '1' HOURS + INTERVAL '1' MINUTES + INTERVAL '1' SECONDS + INTERVAL '1' DAYS + INTERVAL '1' HOURS;
-                   time
-                --------
-                 02:01:00
-                (1 row)
-                
-                SELECT INTERVAL '2' HOURS + '15:12:24'::time;
-                 time
-                ------
-                 17:12:24
-                (1 row)
-                """
-        );
-    }
-
-    @Test
-    public void testTimeSubInterval() {
-        this.qs("""
-                SELECT '00:00:00'::time - INTERVAL '10' MINUTES;
-                   time
-                --------
-                 23:50:00
-                (1 row)
-                
-                SELECT '00:00:00'::time - INTERVAL '-10' MINUTES;
-                   time
-                --------
-                 00:10:00
-                (1 row)
-                
-                SELECT '00:00:00'::time - INTERVAL '2' DAYS;
-                 time
-                --------
-                 00:00:00
-                (1 row)
-                
-                SELECT '00:00:00'::time - INTERVAL '-2' DAYS;
-                 time
-                --------
-                 00:00:00
-                (1 row)
-                
-                SELECT '00:00:00'::time - INTERVAL '2' HOURS;
-                 time
-                --------
-                 22:00:00
-                (1 row)
-                
-                SELECT '00:00:00'::time - INTERVAL '-2' HOURS;
-                 time
-                --------
-                 02:00:00
-                (1 row)
-                
-                SELECT '00:00:00'::time - INTERVAL '2' SECONDS;
-                 time
-                --------
-                 23:59:58
-                (1 row)
-                
-                SELECT '00:00:00'::time - INTERVAL '-2' SECONDS;
-                 time
-                --------
-                 00:00:02
-                (1 row)
-                
-                SELECT '00:00:00'::time - INTERVAL '2' HOURS - INTERVAL '10' MINUTES - INTERVAL '10' SECONDS;
-                 time
-                --------
-                 21:49:50
-                (1 row)
-                
-                SELECT '00:00:00'::time - INTERVAL '-2' HOURS - INTERVAL '-10' MINUTES - INTERVAL '-10' SECONDS;
-                 time
-                --------
-                 02:10:10
-                (1 row)
-                
-                SELECT '00:00:00'::time - INTERVAL '2' HOURS - INTERVAL '10' MINUTES - INTERVAL '-10' SECONDS;
-                 time
-                --------
-                 21:50:10
-                (1 row)
-                """);
-    }
-
-    @Test @Ignore("https://github.com/feldera/feldera/issues/1658")
-    public void testTimeAddInterval1() {
-        this.qs("""
-                SELECT '12:34:56'::time + INTERVAL '25' DAYS;
-                   time
-                ---------
-                 12:34:56
-                (1 row)
-                """
-        );
-    }
-
-    @Test
-    public void testTimeAddInterval2() {
-        this.qs("""
-                SELECT '12:34:56'::time + INTERVAL '1' YEAR;
-                 time
-                ------
-                 12:34:56
-                (1 row)
-                
-                SELECT '12:34:56'::time + INTERVAL '2' YEARS;
-                 time
-                ------
-                 12:34:56
-                (1 row)
-                
-                SELECT '12:34:56'::time + INTERVAL '2' MONTHS;
-                 time
-                ------
-                 12:34:56
-                (1 row)
-                
-                SELECT '12:34:56'::time + INTERVAL '2' MONTH;
-                 time
-                ------
-                 12:34:56
-                (1 row)
-                
-                SELECT '12:34:56'::time + INTERVAL '10 10:30' DAY TO MINUTE;
-                 time
-                ------
-                 23:04:56
-                (1 row)
-                """
-        );
-    }
-
-    @Test
-    public void testTimeSubInterval1() {
-        this.qs("""
-                SELECT '12:34:56'::time - INTERVAL '1' YEAR;
-                 time
-                ------
-                 12:34:56
-                (1 row)
-                
-                SELECT '12:34:56'::time - INTERVAL '2' YEARS;
-                 time
-                ------
-                 12:34:56
-                (1 row)
-                
-                SELECT '12:34:56'::time - INTERVAL '2' MONTHS;
-                 time
-                ------
-                 12:34:56
-                (1 row)
-                
-                SELECT '12:34:56'::time - INTERVAL '2' MONTH;
-                 time
-                ------
-                 12:34:56
-                (1 row)
-                
-                SELECT '12:34:56'::time - INTERVAL '10 10:30' DAY TO MINUTE;
-                 time
-                ------
-                 02:04:56
-                (1 row)
-                """);
     }
 }
